Fully implemented fleet field
Fleet field now correctly transer an amount, based on the number of
fleets owned by the ownining player, from the player who landed on the
field, to the owning player. Fixed up and polished writing and made
several messages to the user more clearly.
Now the displayed balance is also updated for all players after every
turn instead of only the active player